+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## master (unreleased)
 
 ### New features
+* [#395] Added `--show-cops` option to show available cops.
 
 ### Bugs fixed
 
@@ -154,6 +155,7 @@
 * New cop `EndBlock` tracks uses of `END` blocks.
 * New cop `DotPosition` tracks the dot position in multi-line method calls.
 * New cop `Attr` tracks uses of `Module#attr`.
+* Add support for auto-correction of some offences with `-a`/`--auto-correct`.
 
 ### Changes